Add service worker for offline functionality
New service worker files are added, and related logic is inserted into the main.tsx file to register this service worker. Several new dependencies related to Workbox, a set of libraries that power service workers, are added to package.json. This commit prepares the application for offering offline functionality and potential PWA (Progressive Web Application) features in the future.