Update application name and link manifest
The application's short name and full name in the manifest.json file were updated. Also, a link to the manifest.json file was added to the index.html file. The previous names suggested that this was a sample React application, but has since been updated to reflect its true identity as an open-source Tailwind CSS admin dashboard template.