pwa (#1)
* Add service worker for offline functionality

New service worker files are added, and related logic is inserted into the main.tsx file to register this service worker. Several new dependencies related to Workbox, a set of libraries that power service workers, are added to package.json. This commit prepares the application for offering offline functionality and potential PWA (Progressive Web Application) features in the future.

* Update application name and link manifest

The application's short name and full name in the manifest.json file were updated. Also, a link to the manifest.json file was added to the index.html file. The previous names suggested that this was a sample React application, but has since been updated to reflect its true identity as an open-source Tailwind CSS admin dashboard template.

---------

Co-authored-by: Rok Bermež <rok@holosheep,eu>